% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_create_custom_button.R
\name{ro_create_custom_button}
\alias{ro_create_custom_button}
\title{Create a Custom Button for Shiny UI}
\usage{
ro_create_custom_button(input_id, label_text, svg_path)
}
\arguments{
\item{input_id}{Character. The input ID for the action button.}

\item{label_text}{Character. The label to display next to the icon.}

\item{svg_path}{Character. SVG path data for the icon.}
}
\value{
A Shiny tag list representing the action button UI element.
}
\description{
Generates a Shiny action button with an SVG icon and custom label text.
}
\examples{
\dontrun{
ro_create_custom_button(
  input_id = "download_btn",
  label_text = "Download",
  svg_path = ro_icon_download()
)
}
}
\seealso{
\code{\link{ro_icon_download}}, \code{\link{ro_icon_table}}

Other DT:
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
