% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_icon_download.R
\name{ro_icon_download}
\alias{ro_icon_download}
\title{SVG path for download icon}
\usage{
ro_icon_download()
}
\value{
A character string with the SVG path data for a download icon.
}
\description{
SVG path for download icon
}
\examples{
\dontrun{
download_button <- ro_icon_download()
}
}
\seealso{
\code{\link{ro_create_custom_button}}, \code{\link{ro_icon_table}},
\code{\link{ro_shiny_graph_panel_server}}, \code{\link{ro_shiny_graph_panel_ui}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
