% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_download_modules.R
\name{ro_shiny_download_server}
\alias{ro_shiny_download_server}
\title{Download Button Module Server}
\usage{
ro_shiny_download_server(
  id,
  data_to_download,
  filename_prefix = "data",
  svg_path
)
}
\arguments{
\item{id}{Character. The module namespace ID.}

\item{data_to_download}{Function. A function returning the data to be downloaded (as a data frame or tibble).}

\item{filename_prefix}{Character. Prefix for the downloaded file name. Default is "data".}

\item{svg_path}{Character. SVG path data for the icon (not used in server, but included for consistency).}
}
\value{
None. Sets up the server logic for the download button module.
}
\description{
Server logic for a download button that exports data as a CSV file using a custom filename.
}
\examples{
# In server function:
if (interactive()) {
ro_shiny_download_server(
  id = "download1",
  data_to_download = function() mydata,
  filename_prefix = "cases",
  svg_path = ro_icon_download()
)
}
}
\seealso{
\code{\link{ro_shiny_download_ui}}, \code{\link{ro_shiny_graph_panel_server}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
