% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_download_modules.R
\name{ro_shiny_download_ui}
\alias{ro_shiny_download_ui}
\title{Download Button Module UI}
\usage{
ro_shiny_download_ui(id, label_text = "Export", svg_path)
}
\arguments{
\item{id}{Character. The module namespace ID.}

\item{label_text}{Character. The label to display on the button. Default is "Export".}

\item{svg_path}{Character. SVG path data for the icon.}
}
\value{
A Shiny UI element (action button) for use in a module UI.
}
\description{
Creates a Shiny UI element for a download button with a custom SVG icon and label.
}
\examples{
if (interactive()) {
ro_shiny_download_ui(
  id = "download1",
  label_text = "Export",
  svg_path = ro_icon_download()
)
}
}
\seealso{
\code{\link{ro_shiny_download_server}}, \code{\link{ro_shiny_graph_panel_ui}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
