% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_dt_modules.R
\name{ro_shiny_dt_ui}
\alias{ro_shiny_dt_ui}
\title{DT Table Module UI}
\usage{
ro_shiny_dt_ui(id)
}
\arguments{
\item{id}{Character. The module namespace ID.}
}
\value{
A Shiny UI output element for rendering a DT table.
}
\description{
Creates a Shiny UI element for displaying a DT (DataTable) table output within a module.
}
\examples{
if (interactive()) {
ro_shiny_dt_ui("table1")
}
}
\seealso{
\code{\link{ro_shiny_dt_server}}, \code{\link{ro_shiny_graph_panel_server}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
