% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_graph_modules.R
\name{ro_shiny_graph_mod_server}
\alias{ro_shiny_graph_mod_server}
\title{Generic Graph Module Server}
\usage{
ro_shiny_graph_mod_server(id, render_fun, plot_data_fun)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{render_fun}{Function.
A Shiny render function (\code{renderPlotly}, \code{highcharter::renderHighchart},
\code{echarts4r::renderEcharts4r}, etc.).}

\item{plot_data_fun}{Function. A function returning the data or object to plot.}
}
\value{
None. Sets up the server logic for the generic chart module.
}
\description{
Server logic for a generic chart module.
Renders a chart using a user-supplied render function and a plot-producing function.
}
\examples{
# Using plotly
if (interactive()) {
  ro_shiny_graph_mod_server(
    id = "my_chart",
    render_fun = plotly::renderPlotly,
    plot_data_fun = function() plotly_obj
  )
}
# Using highcharter
if (interactive()) {
  ro_shiny_graph_mod_server(
    id = "my_chart",
    render_fun = highcharter::renderHighchart,
    plot_data_fun = function() highcharter::highchart() |>
      highcharter::hc_add_series(data = c(1,2,3))
  )
}

# Using echarts4r
if (interactive()) {
  ro_shiny_graph_mod_server(
    id = "my_chart",
    render_fun = echarts4r::renderEcharts4r,
    plot_data_fun = function() echarts4r::e_charts(1:3) |> echarts4r::e_line(1:3)
  )
}
}
\seealso{
\code{\link{ro_shiny_graph_panel_server}}, \code{\link{ro_shiny_graph_mod_ui}}

Other ggplotly:
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other plotly:
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other echarts:
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}}
}
\concept{echarts}
\concept{ggplotly}
\concept{plotly}
