% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_graph_modules.R
\name{ro_shiny_graph_mod_ui}
\alias{ro_shiny_graph_mod_ui}
\title{Generic yGraph Module UI}
\usage{
ro_shiny_graph_mod_ui(id, plot_ui_fun)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{plot_ui_fun}{Function. A function that generates a UI element for the chart output
(\code{plotlyOutput}, \code{highcharter::highchartOutput}, \code{echarts4r::echarts4rOutput}, etc.).}
}
\value{
A Shiny UI output element for use in a module UI.
}
\description{
Returns a UI output element for a generic chart module, using a user-supplied chart UI function.
}
\examples{
# Using plotly
if (interactive()) {
  ro_shiny_graph_mod_ui("my_chart", plotlyOutput)
}

# Using highcharter
if (interactive()) {
  ro_shiny_graph_mod_ui("my_chart", highcharter::highchartOutput)
}

# Using echarts4r
if (interactive()) {
  ro_shiny_graph_mod_ui("my_chart", echarts4r::echarts4rOutput)
}
}
\seealso{
\code{\link{ro_shiny_graph_panel_ui}}, \code{\link{ro_shiny_graph_mod_server}}

Other ggplotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other plotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other echarts4r:
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}
}
\concept{echarts4r}
\concept{ggplotly}
\concept{plotly}
