% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_graph_panel.R
\name{ro_shiny_graph_panel_server}
\alias{ro_shiny_graph_panel_server}
\title{Graph Panel Module Server}
\usage{
ro_shiny_graph_panel_server(
  id,
  plot_render_fun,
  plot_data_fun,
  data,
  caption,
  sorting = FALSE,
  pagelength = 10
)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{plot_render_fun}{Function. A Shiny render function
(\code{plotly::renderPlotly}, \code{highcharter::renderHighchart}, etc.).}

\item{plot_data_fun}{Function. A function returning the data or plot object for the chart.}

\item{data}{Data frame or tibble. The data to be shown in the table and available for download.}

\item{caption}{Character. Caption/title for the data table.}

\item{sorting}{Logical. Should the table allow column sorting? Default is \code{FALSE}.
When used, apply the same in ro_shiny_graph_panel_ui.}

\item{pagelength}{Integer. Number of rows per page in the data table. Default is \code{10}.}
}
\value{
None. Sets up the server logic for the graph panel module.
}
\description{
Server logic for the graph panel module.
Handles rendering of the chart, toggling and displaying the table, and downloading data.
}
\examples{
# In your Shiny server:
if (interactive()) {
  ro_shiny_graph_panel_server(
  id = "main",
  plot_render_fun = plotly::renderPlotly,
  plot_data_fun = function() plotly_obj,
  data = my_data,
  caption = "Aantal gevallen per groep",
  sorting = TRUE,
  pagelength = 20
  )
}
}
\seealso{
\code{\link{ro_shiny_graph_panel_ui}}, \code{\link{ro_shiny_graph_mod_server}},
\code{\link{ro_show_table_server}}, \code{\link{ro_shiny_download_server}},
\code{\link{ro_icon_download}}, \code{\link{ro_shiny_dt_ui}},
\code{\link{ro_shiny_dt_server}}

Other ggplotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other plotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}}

Other echarts:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
\concept{echarts}
\concept{ggplotly}
\concept{plotly}
