% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_graph_panel.R
\name{ro_shiny_graph_panel_ui}
\alias{ro_shiny_graph_panel_ui}
\title{Graph Panel Module UI}
\usage{
ro_shiny_graph_panel_ui(id, plot_ui_fun, sorting = FALSE)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{plot_ui_fun}{Function. UI function to create the chart output (for example \code{plotlyOutput}).}

\item{sorting}{Logical. Should the table allow column sorting? Default is \code{FALSE}.
When used, apply the same in ro_shiny_graph_panel_server.}
}
\value{
A Shiny UI element (tagList) for use in a module UI.
}
\description{
Generates a Shiny UI module including a chart,
skip-link for accessibility, a table toggle button, download button, and a data table area.
}
\examples{
if (interactive()) {
  ro_shiny_graph_panel_ui("main", plotlyOutput)
}
}
\seealso{
\code{\link{ro_shiny_graph_panel_server}},
\code{\link{ro_shiny_skip_link}}, \code{\link{ro_shiny_skip_link_style}},
\code{\link{ro_shiny_focus_style}}, \code{\link{ro_shiny_button_theme}},
\code{\link{ro_shiny_dt_responsive}}, \code{\link{ro_shiny_screenreader}},
\code{\link{ro_icon_download}}, \code{\link{ro_shiny_graph_mod_ui}},
\code{\link{ro_show_table_ui}}, \code{\link{ro_shiny_download_ui}}

Other ggplotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}}

Other plotly:
\code{\link[=ro_shiny_graph_mod_server]{ro_shiny_graph_mod_server()}},
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}}

Other echarts4r:
\code{\link[=ro_shiny_graph_mod_ui]{ro_shiny_graph_mod_ui()}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_screenreader]{ro_shiny_screenreader()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}
}
\concept{DT}
\concept{echarts4r}
\concept{ggplotly}
\concept{plotly}
