% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ro_shiny_screenreader.R
\name{ro_shiny_screenreader}
\alias{ro_shiny_screenreader}
\title{Screen Reader Only CSS Classes for Shiny Apps}
\usage{
ro_shiny_screenreader()
}
\value{
A \code{shiny.tags} object with a \code{<head>} tag containing screen reader only CSS classes.
}
\description{
Provides general-purpose CSS classes that visually hide content while keeping it
accessible to screen readers. Includes \code{.sr-only} and \code{.visually-hidden} classes.
}
\examples{
if (interactive()) {
  ui <- fluidPage(
    ro_dt_theme(sorting = FALSE),
    ro_shiny_screenreader(),
    DTOutput("mytable")
  )
  server <- function(input, output, session) {
    output$mytable <- renderDT(datatable(iris, options(list(ordering=FALSE))))
  }
  shinyApp(ui, server)
}

}
\seealso{
\code{\link{ro_shiny_skip_link}},
\code{\link{ro_shiny_skip_link_style}}, \code{\link{ro_shiny_dt_ui}},
\code{\link{ro_shiny_graph_panel_ui}}

Other DT:
\code{\link[=ro_create_custom_button]{ro_create_custom_button()}},
\code{\link[=ro_icon_download]{ro_icon_download()}},
\code{\link[=ro_icon_table]{ro_icon_table()}},
\code{\link[=ro_shiny_download_server]{ro_shiny_download_server()}},
\code{\link[=ro_shiny_download_ui]{ro_shiny_download_ui()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_dt_server]{ro_shiny_dt_server()}},
\code{\link[=ro_shiny_dt_ui]{ro_shiny_dt_ui()}},
\code{\link[=ro_shiny_graph_panel_server]{ro_shiny_graph_panel_server()}},
\code{\link[=ro_shiny_graph_panel_ui]{ro_shiny_graph_panel_ui()}},
\code{\link[=ro_show_table_server]{ro_show_table_server()}},
\code{\link[=ro_show_table_ui]{ro_show_table_ui()}}

Other shiny:
\code{\link[=ro_shiny_button_theme]{ro_shiny_button_theme()}},
\code{\link[=ro_shiny_dt_responsive]{ro_shiny_dt_responsive()}},
\code{\link[=ro_shiny_focus_style]{ro_shiny_focus_style()}},
\code{\link[=ro_shiny_skip_link]{ro_shiny_skip_link()}},
\code{\link[=ro_shiny_skip_link_style]{ro_shiny_skip_link_style()}}
}
\concept{DT}
\concept{shiny}
