% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis.shiny-package.R
\docType{package}
\name{ROvis.shiny-package}
\alias{ROvis.shiny}
\alias{ROvis.shiny-package}
\title{ROvis.shiny: Rijksoverheid Visualisatie - shiny}
\description{
ROvis.shiny is an R package that provides a comprehensive suite of utilities for visualizing in Rijksoverheid style.
}
\author{
\strong{Maintainer}: spin@rivm.nl \email{spin@rivm.nl}

Authors:
\itemize{
  \item spin@rivm.nl \email{spin@rivm.nl}
}

}
\keyword{internal}
